+++ src/main/java/teamport/aether/items/itemtool/ItemShooter.java
@@ -1,4 +1,4 @@
-package teamport.aether.items;
+package teamport.aether.items.itemtool;
 
 import net.minecraft.core.entity.player.Player;
 import net.minecraft.core.item.Item;
@@ -6,6 +6,7 @@ import net.minecraft.core.item.ItemStack;
 import net.minecraft.core.world.World;
 import teamport.aether.entity.projectile.ProjectileDartEnchanted;
 import teamport.aether.entity.projectile.ProjectileDart;
+import teamport.aether.items.AetherItems;
 
 public class ItemShooter extends Item {
     public ItemShooter(String name, String namespaceId, int id) {
