+++ src/main/java/teamport/aether/mixin/PlayerControllerValkToolsReachMixin.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import teamport.aether.items.AetherToolMaterial;
+import teamport.aether.items.itemtool.AetherToolMaterial;
 import teamport.aether.mixin.accessors.ItemToolSwordAccessor;
 
 @Mixin(value = PlayerController.class, remap = false)
